Feature/offers (#6)
* add ItemAction

* refactor

* long -> int

* add delete endpoint

* refactor

* fix

* add SERIAL ids, reservations refactor

* fix reservations

* fix

* fix

* add deleteCustomer

* add cors config in application.yml

* refactor

* refactor

* add ResponseData, Filter

* refactor

* add filter param

* add filter functionality

* add pagination to offers

+++ server-app/src/main/java/edu/put/serverapp/repositories/OfferRepository.java
@@ -1,9 +1,9 @@
 package edu.put.serverapp.repositories;
 
-import edu.put.serverapp.models.Offer;
+import edu.put.serverapp.models.entities.Offer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface OfferRepository extends JpaRepository<Offer, Long> {
+public interface OfferRepository extends JpaRepository<Offer, Integer> {
 }
